+++ src/ie/usage_report_srr.rs
@@ -205,7 +205,7 @@ mod tests {
         let max_report = UsageReport::new(
             UrrId::new(u32::MAX),
             SequenceNumber::new(u32::MAX),
-            UsageReportTrigger::new(u8::MAX),
+            UsageReportTrigger::new(u32::from(u8::MAX)),
         );
         let srr = UsageReportSrr::new(max_report);
         let marshaled = srr.marshal();
